Move fetching logic from TotalUsers to Dashboard component; Add loading spinner
